requirements.txt
autograd==1.3
cvxopt==1.2.5
cvxpy==1.1.10
cycler==0.10.0
ecos==2.0.7.post1
future==0.18.2
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.20.1
osqp==0.6.2.post0
pandas==1.2.2
Pillow==8.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
qdldl==0.1.5.post0
scikit-learn==0.24.1
scipy==1.6.0
scs==2.1.2
six==1.15.0
sklearn==0.0
threadpoolctl==2.1.0
torch==1.7.1
torchvision==0.8.2
typing-extensions==3.7.4.3
